Added Hamilton last year, forgot to push it.
Also because I'm paranoid about the finite number of options
for this thing, I'm going to periodically shuffle the owner
labels, which has nothing to do with the random generation,
just the labels that get attached to the final results.  If
the owners get randomly shuffled, that will theoretically add
another layer of randomness?